+++ arm9/source/emucore/AudioOutputLine.h
@@ -14,7 +14,7 @@
 
 #include "types.h"
 
-extern INT64 sampleBuffer[3];
+extern INT32 sampleBuffer[3];
 extern INT32 commonClockCounter[3];
 extern INT32 commonClocksPerSample[3];
 extern INT16 previousSample[3];
